refactor(MFLP-74): Renaming packages
