+++ buyhood-product-category/src/main/java/api/buyhood/category/BuyHoodProductCategoryApplication.java
@@ -1,4 +1,4 @@
-package api.buyhood.productcategory;
+package api.buyhood.category;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
