+++ buyhood-product-category/src/main/java/api/buyhood/category/dto/request/store/PatchStoreCategoryReq.java
@@ -1,4 +1,4 @@
-package api.buyhood.productcategory.dto.request.store;
+package api.buyhood.category.dto.request.store;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
