+++ buyhood-product-category/src/main/java/api/buyhood/category/dto/response/store/GetStoreCategoryRes.java
@@ -1,22 +1,21 @@
-package api.buyhood.productcategory.dto.response.store;
+package api.buyhood.category.dto.response.store;
 
-import api.buyhood.productcategory.entity.StoreCategory;
+import api.buyhood.category.entity.StoreCategory;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public class CreateStoreCategoryRes {
+public class GetStoreCategoryRes {
 
 	private final Long categoryId;
 	private final String categoryName;
 
-	public static CreateStoreCategoryRes of(StoreCategory category) {
-		return new CreateStoreCategoryRes(
+	public static GetStoreCategoryRes of(StoreCategory category) {
+		return new GetStoreCategoryRes(
 			category.getId(),
 			category.getName()
 		);
 	}
-
 }
