+++ buyhood-product-category/src/main/java/api/buyhood/category/repository/StoreCategoryRepository.java
@@ -1,6 +1,6 @@
-package api.buyhood.productcategory.repository;
+package api.buyhood.category.repository;
 
-import api.buyhood.productcategory.entity.StoreCategory;
+import api.buyhood.category.entity.StoreCategory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
