+++ buyhood-product-category/src/test/java/api/buyhood/category/BuyHoodCategoryApplicationTests.java
@@ -1,4 +1,4 @@
-package api.buyhood.productcategory;
+package api.buyhood.category;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
